refactor: move admin feature flags to featureFlags/admin directory
