+++ src/schema/v2/schema.ts
@@ -131,11 +131,11 @@ import { TargetSupply } from "./TargetSupply"
 import { UserField } from "./user"
 // import TrendingArtists from "./artists/trending"
 import { updateQuizMutation } from "schema/v2/updateQuizMutation"
-import { AdminField } from "./admin"
-import { createFeatureFlagMutation } from "./admin/mutations/createFeatureFlagMutation"
-import { deleteFeatureFlagMutation } from "./admin/mutations/deleteFeatureFlagMutation"
-import { toggleFeatureFlagMutation } from "./admin/mutations/toggleFeatureFlagMutation"
-import { updateFeatureFlagMutation } from "./admin/mutations/updateFeatureFlagMutation"
+import { AdminField } from "./featureFlags/admin"
+import { createFeatureFlagMutation } from "./featureFlags/admin/mutations/createFeatureFlagMutation"
+import { deleteFeatureFlagMutation } from "./featureFlags/admin/mutations/deleteFeatureFlagMutation"
+import { toggleFeatureFlagMutation } from "./featureFlags/admin/mutations/toggleFeatureFlagMutation"
+import { updateFeatureFlagMutation } from "./featureFlags/admin/mutations/updateFeatureFlagMutation"
 import { channel } from "./article/channel"
 import { createArtistMutation } from "./artist/createArtistMutation"
 import { deleteArtistMutation } from "./artist/deleteArtistMutation"
